Extract SampleItem
